Turbidity Output Data
Hi everyone!
ActiveLearning data are available on GitHub/Output: we've redone the plates for turbidity measurement and data are available in the file “Output /2023.10.25”.
To get an idea of the robot's reproducibility and precision (data are in the “Discrepancy_Test_Robot_HandMade” file)
Some cases from the original file “2023.10.25-JS.xlsx” (see last sheet of "Discrepancy_test_Composition.xlsx") have been prepared twice by hand and twice using the robot in triplicates.
We want to have an idea of the discrepancy in turbidity between triplets at the same location on the same plate.
The discrepancy in turbidity between the 2 triplicates on the same plate.
The discrepancy from one plate to another. This will be done between plate 1 and plate 2 made by hand, and those made by the robot.
We then want to compare how accurately the robot is reproducing approximately the same sample versus the human being: we'll need to compare the two hand-made plates to those made by the robot.
The plate composition is displayed on the last sheet of "Discrepancy_test_Composition.xlsx"
We have the same plate made 4 times: 2 by hand and 2 by the robot.